refactor airflow example
